Fixed compilation timeout
